Rename simulation to nlisim
