+++ setup.cfg
@@ -3,6 +3,6 @@ ignore_missing_imports = True
 
 [coverage:run]
 branch = True
-include = simulation/*
-          {envsitepackagesdir}/simulation/*
+include = nlisim/*
+          {envsitepackagesdir}/nlisim/*
 omit = test/*
