+++ setup.py
@@ -2,7 +2,7 @@
 from setuptools import find_packages, setup
 
 setup(
-    name='simulation',
+    name='nlisim',
     packages=find_packages(exclude=['test', 'test.*']),
     install_requires=[
         'attrs',
@@ -15,5 +15,5 @@ setup(
         'tqdm',
         'vtk',
     ],
-    entry_points={'console_scripts': ['simulation = simulation.cli:main']},
+    entry_points={'console_scripts': ['nlisim = nlisim.cli:main']},
 )
