+++ test/test_config.py
@@ -3,8 +3,8 @@ from tempfile import NamedTemporaryFile
 
 from pytest import raises
 
-from simulation.config import SimulationConfig
-from simulation.module import Module
+from nlisim.config import SimulationConfig
+from nlisim.module import Module
 
 
 def test_config_defaults():
@@ -56,7 +56,7 @@ def test_config_multiple():
 
 def test_config_add_module_string():
     config = SimulationConfig()
-    config.add_module('simulation.modules.afumigatus.Afumigatus')
+    config.add_module('nlisim.modules.afumigatus.Afumigatus')
     assert len(config.modules) == 1
     assert isinstance(config.modules[0], Module)
