+++ tox.ini
@@ -42,19 +42,19 @@ skip_install = true
 deps =
     mypy
 commands =
-    mypy {posargs} simulation
+    mypy {posargs} nlisim
 
 [testenv:docs]
 deps =
     pdoc3
-commands = pdoc3 -o docs --config latex_math=True --force --html simulation
+commands = pdoc3 -o docs --config latex_math=True --force --html nlisim
 
 [flake8]
 max-line-length = 100
 show-source = True
 # https://github.com/PyCQA/flake8-import-order
 import-order-style = appnexus
-application-import-names = simulation
+application-import-names = nlisim
 ignore =
     # missing whitespace after ',' (conflicts with black's formatting rules)
     E231
